properly update paths on drag and fix warnings

+++ src/DOM/CharacterDOMNode.cpp
@@ -193,6 +193,6 @@ void LCharacterDOMNode::PreProcess()
 		if (index == -1)
 			mItemTableIndex = 0;
 		else
-			mItemTableIndex = index;
+			mItemTableIndex = static_cast<int32_t>(index);
 	}
 }
