+++ src/DOM/FurnitureDOMNode.cpp
@@ -274,6 +274,6 @@ void LFurnitureDOMNode::PreProcess()
 		if (index == -1)
 			mItemTableIndex = 0;
 		else
-			mItemTableIndex = index;
+			mItemTableIndex = static_cast<int32_t>(index);
 	}
 }
